core: services: commander: Do not delete folders and open files

+++ core/services/commander/main.py
@@ -1,7 +1,6 @@
 #! /usr/bin/env python3
 import logging
 import os
-import shutil
 import time
 from enum import Enum
 from pathlib import Path
@@ -11,6 +10,7 @@ import appdirs
 import uvicorn
 from commonwealth.utils.apis import GenericErrorHandlingRoute
 from commonwealth.utils.commands import run_command
+from commonwealth.utils.general import file_is_open
 from commonwealth.utils.logs import InterceptHandler, get_new_log_path
 from fastapi import FastAPI, HTTPException, status
 from fastapi.responses import HTMLResponse
@@ -49,17 +49,17 @@ def check_what_i_am_doing(i_know_what_i_am_doing: bool = False) -> None:
 
 
 def delete_everything(path: Path) -> None:
-    if path.is_file():
+    if path.is_file() and not file_is_open(path):
         path.unlink()
         return
 
     for item in path.glob("*"):
         try:
-            if item.is_file():
+            if item.is_file() and not file_is_open(item):
                 item.unlink()
-            if item.is_dir():
-                # Delete folder and its contents
-                shutil.rmtree(item)
+            if item.is_dir() and not item.is_symlink():
+                # Delete folder contents
+                delete_everything(item)
         except Exception as exception:
             logger.warning(f"Failed to delete: {item}, {exception}")
